requirements.txt
torch
torchvision
pytest
matplotlib
pillow
numpy